+++ src/main/java/me/mrCookieSlime/QuestWorld/quest/Quest.java
@@ -26,7 +26,7 @@ class Quest extends Renderable implements IQuestState {
 	
 	WeakReference<Category> category;
 	int id;
-	int cooldown;
+	long cooldown;
 	String name;
 	ItemStack item = new ItemStack(Material.BOOK_AND_QUILL);
 	List<Mission> tasks = new ArrayList<>();
@@ -83,11 +83,11 @@ class Quest extends Renderable implements IQuestState {
 		dest.copy(this);
 	}
 	
-	private int fromMaybeString(Object o) {
-		if(o instanceof Integer)
-			return ((Integer)o).intValue();
+	private long fromMaybeString(Object o) {
+		if(o instanceof Long)
+			return ((Long)o).longValue();
 		if(o instanceof String)
-			return Integer.valueOf((String)o);
+			return Long.valueOf((String)o);
 		
 		throw new IllegalArgumentException("Expected Integer or String, got " + o.getClass().getSimpleName());
 	}
@@ -307,19 +307,19 @@ class Quest extends Renderable implements IQuestState {
 		partysize = size;
 	}
 
-	public int getRawCooldown() {
+	public long getRawCooldown() {
 		return cooldown;
 	}
 	
-	public void setRawCooldown(int cooldown) {
+	public void setRawCooldown(long cooldown) {
 		this.cooldown = cooldown;
 	}
 	
-	public int getCooldown() {
+	public long getCooldown() {
 		return cooldown / 60 / 1000;
 	}
 
-	public void setCooldown(int cooldown) {
+	public void setCooldown(long cooldown) {
 		this.cooldown = cooldown * 60 * 1000;
 	}
